+++ tools/content-production/index.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { ContentProductionTool } from "./ContentProductionTool";
+import { config } from "./config";
+
+export default function ContentProductionToolPage() {
+  return <ContentProductionTool config={config} />;
+}